+++ sain/iter.py
@@ -48,6 +48,8 @@ __all__ = (
     "Chunks",
     "Empty",
     "Repeat",
+    "Once",
+    "ExactSizeIterator",
     # Functions
     "into_iter",
     "empty",
@@ -65,7 +67,7 @@ from . import default as _default
 from . import futures
 from . import option as _option
 from . import result as _result
-from .collections import vec
+from .collections import vec as _vec
 from .macros import rustc_diagnostic_item
 
 Item = typing.TypeVar("Item")
@@ -173,11 +175,9 @@ class Iterator(
         return Empty()
 
     @typing.overload
-    @typing.final
     def collect(self) -> collections.Sequence[Item]: ...
 
     @typing.overload
-    @typing.final
     def collect(
         self, *, cast: collections.Callable[[Item], OtherItem]
     ) -> collections.Sequence[OtherItem]: ...
@@ -236,7 +236,7 @@ class Iterator(
                 collection[idx] = item
 
     @typing.final
-    def to_vec(self) -> vec.Vec[Item]:
+    def to_vec(self) -> _vec.Vec[Item]:
         """Convert this iterator into `Vec[T]`.
 
         Example
@@ -248,7 +248,7 @@ class Iterator(
         assert to_vec == [0]
         ```
         """
-        return vec.Vec(_ for _ in self)
+        return _vec.Vec(_ for _ in self)
 
     @typing.final
     def sink(self) -> None:
@@ -553,7 +553,7 @@ class Iterator(
 
     def zip(
         self, other: collections.Iterable[OtherItem]
-    ) -> Iterator[tuple[Item, OtherItem]]:
+    ) -> Iter[tuple[Item, OtherItem]]:
         """Zips the iterator with another iterable.
 
         Example
@@ -583,7 +583,7 @@ class Iterator(
         *,
         key: collections.Callable[[Item], _typeshed.SupportsRichComparison],
         reverse: bool = False,
-    ) -> Iterator[Item]:
+    ) -> Iter[Item]:
         """Sorts the iterator.
 
         Example
@@ -608,7 +608,7 @@ class Iterator(
         """
         return Iter(sorted(self.raw_parts(), key=key, reverse=reverse))
 
-    def reversed(self) -> Iterator[Item]:
+    def reversed(self) -> Iter[Item]:
         """Returns a new iterator that yields the items in the iterator in reverse order.
 
         Example
@@ -627,7 +627,7 @@ class Iterator(
         # first collect it into some collection.
         return Iter(reversed(self.collect()))
 
-    def union(self, other: collections.Iterable[Item]) -> Iterator[Item]:
+    def union(self, other: collections.Iterable[Item]) -> Iter[Item]:
         """Returns a new iterator that yields all items from both iterators.
 
         Example
@@ -834,7 +834,7 @@ class Iterator(
         """
         return await futures.spawn(*(func(item) for item in self))
 
-    def __reversed__(self) -> Iterator[Item]:
+    def __reversed__(self) -> Iter[Item]:
         return self.reversed()
 
     def __setitem__(self, _: None, __: None) -> typing.NoReturn:
@@ -856,6 +856,76 @@ class Iterator(
 
     def __iter__(self) -> Iterator[Item]:
         return self
+
+
+class ExactSizeIterator(typing.Generic[Item], Iterator[Item]):
+    """An iterator that knows its exact size.
+
+    The implementations of this interface indicates that the iterator knows exactly
+    how many items it can yield.
+
+    however, this is not a requirement for the iterator to implement this trait, as its
+    only used for iterators that can know their size.
+
+    Example
+    -------
+    ```py
+    @dataclass
+    class Letters(ExactSizeIterator[str]):
+        letters: list[str]
+
+        def __next__(self) -> str:
+            return self.letters.pop(0)
+
+        def __len__(self) -> int:
+            return len(self.letters)
+
+    letters = Letters(['a', 'b', 'c'])
+    assert letters.count() == 3
+    assert letters.next() == Some('a')
+    assert letters.count() == 2
+    ```
+    """
+
+    @typing.final
+    def count(self) -> int:
+        return len(self)
+
+    @typing.final
+    def len(self) -> int:
+        """Returns the length of the iterator.
+
+        This doesn't exhaust the iterator.
+
+        Example
+        -------
+        ```py
+        it = once(0)
+        assert it.len() == 1
+        assert it.len() == 1
+        it.next()
+        assert it.len() == 0
+        ```
+        """
+        return len(self)
+
+    @typing.final
+    def is_empty(self) -> bool:
+        """Return `True` if this iterator has no items left to yield.
+
+        Example
+        -------
+        ```py
+        iterator = once(1)
+        assert not iterator.is_empty()
+        assert once.next() == Some(1)
+        assert iterator.is_empty()
+        ```
+        """
+        return len(self) == 0
+
+    def __len__(self) -> int:
+        return 0
 
 
 @typing.final
@@ -917,10 +987,7 @@ class Iter(Iterator[Item]):
         return Iter(copy.copy(self._it))
 
     def __next__(self) -> Item:
-        try:
-            return next(self._it)
-        except StopIteration:
-            raise
+        return next(self._it)
 
     def __getitem__(self, index: int) -> Option[Item]:
         try:
@@ -1165,7 +1232,7 @@ class Chunks(typing.Generic[Item], Iterator[collections.Sequence[Item]]):
 
 @typing.final
 @diagnostic
-class Empty(typing.Generic[Item], Iterator[Item]):
+class Empty(typing.Generic[Item], ExactSizeIterator[Item]):
     """An iterator that yields nothing.
 
     This is the default iterator that is created by `Iterator.default()` or `empty()`
@@ -1181,7 +1248,7 @@ class Empty(typing.Generic[Item], Iterator[Item]):
         # also we avoid calling `nothing_unchecked()` here for fast returns.
         return _option.NOTHING  # pyright: ignore
 
-    def count(self) -> typing.Literal[0]:
+    def __len__(self) -> typing.Literal[0]:
         return 0
 
     def any(
@@ -1200,7 +1267,7 @@ class Empty(typing.Generic[Item], Iterator[Item]):
 
 @typing.final
 @diagnostic
-class Repeat(typing.Generic[Item], Iterator[Item]):
+class Repeat(typing.Generic[Item], ExactSizeIterator[Item]):
     """An iterator that repeats a given value an exact number of times.
 
     This iterator is created by calling `repeat()`.
@@ -1223,12 +1290,28 @@ class Repeat(typing.Generic[Item], Iterator[Item]):
 
         unreachable()
 
-    # Specializing these methods to avoid unnecessary function calls.
-    def count(self) -> int:
-        return self._count
-
     def __len__(self) -> int:
         return self._count
+
+
+@typing.final
+@diagnostic
+class Once(typing.Generic[Item], ExactSizeIterator[Item]):
+    """An iterator that yields exactly one item.
+
+    This iterator is created by calling `once()`.
+    """
+
+    __slots__ = ("_item",)
+
+    def __init__(self, item: Item) -> None:
+        self.item: Option[Item] = _option.Some(item)
+
+    def __next__(self) -> Item:
+        return self.item.take().unwrap_or_else(unreachable)
+
+    def __len__(self) -> typing.Literal[0, 1]:
+        return 1 if self.item else 0
 
 
 # a hack to trick the type-checker into thinking that this iterator yield `Item`.
@@ -1272,7 +1355,7 @@ def repeat(element: Item, count: int) -> Repeat[Item]:
 
 
 @rustc_diagnostic_item("once")
-def once(item: Item) -> Iterator[Item]:
+def once(item: Item) -> Once[Item]:
     """Returns an iterator that yields exactly a single item.
 
     Example
@@ -1283,13 +1366,13 @@ def once(item: Item) -> Iterator[Item]:
     assert iterator.next() == Some(None)
     ```
     """
-    return Iter((item,))
+    return Once(item)
 
 
 @rustc_diagnostic_item("into_iter")
 def into_iter(
     iterable: collections.Iterable[Item],
-) -> Iterator[Item]:
+) -> Iter[Item]:
     """Convert any iterable into `Iterator[Item]`.
 
     Example