+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Added
+
+- `iter.Once`
+- `iter.ExactSizeIterator`
+
+### Changed
+
+- `iter.once` now returns a specialized iterator.
+
 ## 1.3.0 - 01/01/2025
 
 ### Added